move none page component from pages to srcs
