+++ src/frontend/srcs/components/chatpage/ChatBody.tsx
@@ -3,7 +3,7 @@ import ysungwonIcon from "../../../assets/ysungwon.jpg";
 import Image from "next/image";
 // import { SocketContext } from "@/context/socket";
 // import { SocketContext } from "@/context/socket";
-import axiosApi from "../../../srcs/AxiosInterceptor";
+import axiosApi from "../../AxiosInterceptor";
 const ChatBody = ({
   messages,
   blocklist,
