+++ src/frontend/srcs/components/chatpage/ChatFooter.tsx
