+++ src/frontend/srcs/components/gamelistpage/GameList.tsx
