+++ src/frontend/srcs/components/gamelistpage/GameListHeader.tsx
