+++ src/frontend/srcs/components/gamelistpage/GameListItemInfo.tsx
