+++ src/frontend/srcs/components/modalpage/ModalAlert.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState, useContext } from "react";
 import { SocketContext } from "@/context/socket";
-import "../../index.css";
+import "../../../pages/index.css";
 
 export default function ModalAlert({
   setIsOpenModal,
