+++ src/frontend/srcs/components/navpage/NavBarLogo.tsx
