+++ src/frontend/srcs/components/searchlistpage/SearchList.tsx
