+++ src/frontend/srcs/components/searchlistpage/SearchResult.tsx
